+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Added the ability to resize the monaco editor vertically
+
 ## [0.27.0] - 2024-09-25
 ### Added
 - Added the ability to view task logs using Grafana only for the admin(Logs tab)