+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Lineage graph: redirect to another dataset
 - Application search: redirect to another dataset
 - Lineage tab displayed correctly
+- Prettified histopy timeline
 
 ## [0.27.0] - 2024-09-25
 ### Added